perf: scope DOS path parsing to Windows
