+++ crates/rspack_paths/src/lib.rs
@@ -22,27 +22,6 @@ use rspack_intern::{InternSliceStorage, InternedSlice, SliceInternable};
 use rustc_hash::FxHasher;
 pub use ustr::IdentityHasher;
 
-/// Returns the byte index immediately after a Windows DOS device path prefix
-/// (`\\\\?\\` or `\\\\.\\`), or zero when `path` has no such prefix.
-///
-/// The prefix is ASCII, so the byte index is also a valid UTF-8 slice boundary.
-/// This follows enhanced-resolve's DOS device path handling:
-/// <https://github.com/webpack/enhanced-resolve/pull/551>.
-#[inline]
-pub fn windows_dos_device_path_prefix_len(path: &str) -> usize {
-  let bytes = path.as_bytes();
-  if bytes.len() >= 4
-    && bytes[0] == b'\\'
-    && bytes[1] == b'\\'
-    && matches!(bytes[2], b'?' | b'.')
-    && bytes[3] == b'\\'
-  {
-    4
-  } else {
-    0
-  }
-}
-
 pub trait AssertUtf8 {
   type Output;
   fn assert_utf8(self) -> Self::Output;
