+++ tests/rspack-test/configCases/loader/dos-device-resource/test.filter.js
@@ -1,0 +1,1 @@
+module.exports = () => process.platform === 'win32';
